gateway/requirements.txt
fastapi
uvicorn
requests
python-dotenv
boto3
